+++ lib/src/widgets/selectable.dart
@@ -290,6 +290,8 @@ class SelectableMath extends StatelessWidget {
                   effectiveTextStyle.fontWeight != null
               ? FontOptions(fontWeight: effectiveTextStyle.fontWeight!)
               : null,
+          textModeTextStyle: effectiveTextStyle,
+          textLocale: Localizations.maybeLocaleOf(context),
           logicalPpi: logicalPpi,
           color: effectiveColor,
         );